pyproject.toml
[tool.poetry]
name = "high-order-layers-torch"
version = "1.1.16"
description = "High order layers in pytorch"
authors = ["jloverich <john.loverich@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
torchmetrics = "^0.6.1"
torch = "^1.12.0"
pytorch-lightning = "^1.6.5"
torch-optimizer = "^0.3.0"
hydra-core = "^1.2.0"

[tool.poetry.dev-dependencies]
coverage = "^6.3.3"
black = "^22.3.0"
lightning-bolts = "^0.5.0"
pytest = "^7.1.2"
torchvision = "^0.13.0"
matplotlib = "^3.5.2"
pre-commit = "^2.20.0"
isort = "^5.10.1"
pylint = "^2.15.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
